package/iputils: Reenable traceroute6
Bump to s20190515 removed traceroute6 build (it's not built by default
since this version in upstream). Reenable it again.

Fixes: 9ffcd9279e ("package/iputils: bump to version s20190515")

This requires fix similar to one in 0003-meson.build-fix-build-with-NLS.patch [1]

traceroute6 is in /usr/sbin, so no need to handle busybox problem
(6e18892733 "package/iputils: move binaries to the location also used by
Busybox").

[1] http://autobuild.buildroot.org/results/0a8a3efe734ac7fb3a68ba505277681857dc0a3d

Suggested-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>
Signed-off-by: Petr Vorel <petr.vorel@gmail.com>